% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat_utils.R
\name{allow_subset_stat}
\alias{allow_subset_stat}
\title{Include subset in the Stat}
\usage{
allow_subset_stat(new_stat, old_stat)
}
\arguments{
\item{new_stat}{The name of the new Stat}

\item{old_stat}{The name of the old Stat}
}
\description{
allow subset to be used in the 'aes'
}
\author{
Ruizhu Huang
}
