% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_align_to_ggtree.R
\name{.align_to_ggtree}
\alias{.align_to_ggtree}
\title{align a matrix to a ggtree}
\usage{
.align_to_ggtree(data, data_tree, rel_width = 1)
}
\arguments{
\item{data}{a matrix. Its row names can be mapped to nodes of a tree.}

\item{data_tree}{a data frame that is generated by
\code{\link[ggtree:ggtree]{ggtree}}}

\item{rel_width}{a numeric value. The relative width of a heatmap to the
width the ggtree.}
}
\value{
a data frame
}
\description{
It matches rows of a matrix to nodes of a ggtree by the row name.
}
\author{
Ruizhu Huang
}
\keyword{internal}
