% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tree_utilis.R
\name{.matTree}
\alias{.matTree}
\title{Transform a phylo object into a matrix.}
\usage{
.matTree(tree)
}
\arguments{
\item{tree}{A phylo object}
}
\value{
A matrix
}
\description{
\code{matTree} transforms a phylo tree into a matrix. The entry of the matrix
is node number. Each row represents a path connecting a leaf node and the
root. The columns are arranged in the order as the path passing the nodes to
reach the root.
}
\author{
Ruizhu Huang
}
\keyword{internal}
