% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{.transNode}
\alias{.transNode}
\title{Transfer between node number and node label}
\usage{
.transNode(tree, node, use.alias = FALSE, message = FALSE)
}
\arguments{
\item{tree}{A phylo object}

\item{node}{A character or numeric vector representing tree node label(s) or
tree node number(s)}

\item{use.alias}{A logical value, TRUE or FALSE. This is an optional argument
that only requried when the input \code{node} is a numeric vector. The
default is FALSE, and the node label would be returned; otherwise, the
alias of node label would be output. The alias of node label is created by
adding a prefix \code{"alias_"} to the node number.}

\item{message}{A logical value, TRUE or FALSE. The default is FALSE. If TRUE,
message will show when a tree have duplicated labels for some internal
nodes.}
}
\value{
a vector
}
\description{
\code{transNode} does the transformation between the number and the label of
a node on a tree
}
\author{
Ruizhu Huang
}
\keyword{internal}
