% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_th_axis.R
\name{geom_th_axis}
\alias{geom_th_axis}
\title{add the axis and the border
add the axis and the border of an annotation plot}
\usage{
geom_th_axis(
  name = NULL,
  th_data = NULL,
  axis_ticks_length = NULL,
  axis_color = "black",
  axis_label_color = "black",
  axis_label_size = 3,
  axis_label_nudge_x = 0,
  axis_label_nudge_y = 0,
  axis_direction = "h",
  axis_location = NULL,
  border_color = "grey",
  border_size = 0.5
)
}
\arguments{
\item{name}{NULL or a name of heatmap (\link{geom_th_heatmap}).}

\item{th_data}{a data frame. To draw a vertical axis, it requires columns
\code{axis_minX}, \code{axis_maxX}, \code{axisY} and \code{label}, where
\code{axis_minX} and \code{axis_maxX} decide the border limit on the x-dim.
The axis is either on \code{axis_minX} for \code{axis_location = "left"} or
\code{axis_maxX} for \code{axis_location = "right"}. The axis ticks are
decided by the \code{axisY} and the tick labels are given in \code{label}.
Similarly, to draw a horizontal axis, it requires columns \code{axisX},
\code{axis_minY}, \code{axis_maxY}, and \code{label}.}

\item{axis_ticks_length}{a number to decide the lenght of ticks.}

\item{axis_color}{the color of the axis}

\item{axis_label_color}{the color of axis labels}

\item{axis_label_size}{the size of axis labels}

\item{axis_label_nudge_x}{adjust the location of labels along x-axis}

\item{axis_label_nudge_y}{adjust the location of labels along y-axis}

\item{axis_direction}{either "h" for horizontal or "v" for vertical}

\item{axis_location}{NULL. If \code{axis_direction = "h"}, it should be "top"
or "bottom"; otherwise, it should be "left" or "right".}

\item{border_color}{the color of the rectangular border}

\item{border_size}{the size of the border line}
}
\description{
add the axis and the border
add the axis and the border of an annotation plot
}
\examples{
library(ggplot2)
df_v <- data.frame(axis_minX = rep(1, 10),
                   axis_maxX = rep(10, 10),
                   axisY = 1:10,
                   label = LETTERS[1:10])
ggplot() +
    geom_th_axis(th_data = df_v,
    axis_direction = "v",
    axis_location = "right")


df_h <- data.frame(axisX = 1:10,
                   axis_minY = rep(1, 10),
                   axis_maxY = rep(10, 10),
                   label = LETTERS[1:10])


ggplot() +
    geom_th_axis(th_data = df_h, axis_direction = "h")
}
\author{
Ruizhu Huang
}
