% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_th_bar.R
\name{geom_th_bar}
\alias{geom_th_bar}
\alias{geom_th_bar0}
\title{Annotation barplot}
\usage{
geom_th_bar(
  name = NULL,
  th_data = NULL,
  mapping = NULL,
  side = c("top", "bottom", "right", "left"),
  gap = 1,
  width = 0.3,
  rel_width = 1,
  rel_height = 1,
  n = 2,
  ...,
  axis_ticks_length = NULL,
  axis_color = "black",
  axis_label_color = "black",
  axis_label_size = 3,
  axis_label_nudge_x = 0,
  axis_label_nudge_y = 0,
  axis_location = NULL,
  border_color = "grey",
  border_size = 0.5
)

geom_th_bar0(
  name = NULL,
  th_data = NULL,
  mapping = NULL,
  side = "right",
  gap = 1,
  rel_width = 1,
  rel_height = 1,
  n = 2,
  ...
)
}
\arguments{
\item{name}{a name to select heatmap}

\item{th_data}{a data frame. It should include one column named as
\code{rowLab} to store the row name of the heatmap when \code{side} is
\strong{left} or \strong{right}; otherwise, it should include one column
named as \code{colLab} to store the column name of the heatmap when
\code{side} is \strong{top} or \strong{bottom}.}

\item{mapping}{Set of aesthetic mappings created by
\code{\link[ggplot2]{aes}}.}

\item{side}{a character to specify where to put the barplot. It is selected
from \strong{left}, \strong{right} to annotate rows of the heatmap, or
\strong{top}, \strong{bottom} to annotate columns of the heatmap.}

\item{gap}{a numeric value to specify the gap between the selected heatmap
and the bar plot.}

\item{width}{Bar width. By default, set to 90\% of the resolution of the data.}

\item{rel_width}{a numeric value decide the relative width of the barplot
compared to the selected heatmap}

\item{rel_height}{a numeric value decide the relative height of the barplot
compared to the selected heatmap}

\item{n}{a number to decide the number of ticks. See
\code{\link[base]{pretty}}.}

\item{...}{More arguments accepted by \code{\link[ggplot2]{geom_bar}}.}

\item{axis_ticks_length}{a number to decide the length of ticks.}

\item{axis_color}{the color of the axis}

\item{axis_label_color}{the color of axis labels}

\item{axis_label_size}{the size of axis labels}

\item{axis_label_nudge_x}{adjust the location of labels along \code{x-axis}}

\item{axis_label_nudge_y}{adjust the location of labels along \code{y-axis}}

\item{axis_location}{NULL. If \code{side} is "top" or "bottom", it should be
"left" or "right"; otherwise, it should be "top" or "bottom".}

\item{border_color}{the color of the rectangular border}

\item{border_size}{the size of the border line}
}
\value{
a list of geom layers
}
\description{
This adds a barplot to annotate rows or columns of a heatmap with showing
the scale of the barplot.
}
\author{
Ruizhu Huang
}
