% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_th_bar.R
\name{geom_th_bar0}
\alias{geom_th_bar0}
\title{Annotation barplot}
\usage{
geom_th_bar0(
  name = NULL,
  th_data = NULL,
  mapping = NULL,
  side = "right",
  gap = 1,
  rel_width = 1,
  rel_height = 1,
  stat = "identity",
  n = 2,
  ...,
  inherit.aes = FALSE
)
}
\arguments{
\item{name}{a name to select heatmap}

\item{th_data}{a data frame. It should include one column named as
\code{rowLab} to store the row name of the heatmap when \code{side} is
\strong{left} or \strong{right}; otherwise, it should include one column
named as \code{colLab} to store the column name of the heatmap when
\code{side} is \strong{top} or \strong{bottom}.}

\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{side}{a character to specify where to put the barplot. It is selected
from \strong{left}, \strong{right} to annotate rows of the heatmap, or
\strong{top}, \strong{bottom} to annotate columns of the heatmap.}

\item{gap}{a numeric value to specify the gap between the selected heatmap
and the barplot}

\item{rel_width}{a numeric value decide the relative width of the barplot
compared to the selected heatmap}

\item{rel_height}{a numeric value decide the relative height of the barplot
compared to the selected heatmap}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{n}{a number to decide the number of ticks. See \link[base]{pretty}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\value{
a geom layer
}
\description{
This adds a barplot to annotate rows or columns of a heatmap without showing
the scale of the barplot.
}
\author{
Ruizhu Huang
}
