% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_th_heatmap.R
\name{geom_th_heatmap}
\alias{geom_th_heatmap}
\title{Create heatmaps}
\usage{
geom_th_heatmap(
  name = NULL,
  th_data = NULL,
  cluster_column = FALSE,
  hclust_method = "complete",
  dist_method = "euclidean",
  gap = 1,
  rel_width = 1,
  show_coltree = TRUE,
  gap_coltree = 0.5,
  rel_height_coltree = 0.1,
  color_coltree = "grey",
  size_coltree = 1,
  ...
)
}
\arguments{
\item{name}{the name of the current heatmap}

\item{th_data}{a matrix}

\item{cluster_column}{a logical value, TRUE or FALSE. If TRUE, columns of the
heatmap are rearranged according to their similarities.}

\item{hclust_method}{the method used to do clustering. This should be
"ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median"
or "centroid". See \code{method} for \code{\link[stats]{hclust}}.}

\item{dist_method}{the method used to do clustering. This must be one of
"euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski".
See \code{method} for \code{\link[stats]{dist}}.}

\item{gap}{a numeric value to specify the gap between the current and the
previous heatmap}

\item{rel_width}{a numeric value decide the relative width of the heatmap
compared to the ggtree plot}

\item{show_coltree}{TRUE or FALSE. If TRUE, the column tree is shown when
\code{cluster_column = TRUE}}

\item{gap_coltree}{a numeric value to decide the gap between the heatmap and
the column tree}

\item{rel_height_coltree}{the relative height of the column tree compared to
the height of row tree.}

\item{color_coltree}{the color of the column tree.}

\item{size_coltree}{the line size of the column tree}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{geom_tile}} to
customize the heatmap.}
}
\value{
geom layer
}
\description{
Create heatmaps
}
\author{
Ruizhu Huang
}
