% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_th_heatmap.R
\name{geom_th_heatmap0}
\alias{geom_th_heatmap0}
\title{Create heatmaps}
\usage{
geom_th_heatmap0(
  th_data = NULL,
  cluster_column = FALSE,
  hclust_method = "complete",
  dist_method = "euclidean",
  gap = 1,
  name = NULL,
  rel_width = 1,
  ...,
  linejoin = "mitre",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE
)
}
\arguments{
\item{th_data}{a matrix}

\item{cluster_column}{a logical value, TRUE or FALSE. If TRUE, columns of the
heatmap are rearranged according to their similarities.}

\item{hclust_method}{the method used to do clustering. This should be
"ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median"
or "centroid". See \code{method} for \code{\link[stats]{hclust}}.}

\item{dist_method}{the method used to do clustering. This must be one of
"euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski".
See \code{method} for \code{\link[stats]{dist}}.}

\item{gap}{a numeric value to specify the gap between the current and the
previous heatmap}

\item{name}{the name of the current heatmap}

\item{rel_width}{a numeric value decide the relative width of the heatmap
compared to the ggtree plot}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{linejoin}{Line join style (round, mitre, bevel).}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\value{
geom layer
}
\description{
Create heatmaps
}
\author{
Ruizhu Huang
}
