% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_th_heatmp.R
\name{geom_th_heatmp}
\alias{geom_th_heatmp}
\title{Create heatmaps}
\usage{
geom_th_heatmp(
  mapping = NULL,
  th_data = NULL,
  cluster_column = FALSE,
  hclust_method = "complete",
  dist_method = "euclidean",
  data = NULL,
  position = "identity",
  gap = 1,
  name = NULL,
  rel_width = 1,
  ...,
  linejoin = "mitre",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{th_data}{a matrix}

\item{cluster_column}{a logical value, TRUE or FALSE. If TRUE, columns of the
heatmap are rearranged according to their similarities.}

\item{hclust_method}{the method used to do clustering. See \code{method} for
\code{\link[stats]{hclust}}.}

\item{dist_method}{the method used to do clustering. See \code{method} for
\code{\link[stats]{dist}}.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{gap}{a numeric value to specify the gap between the current and the
previous heatmap}

\item{name}{the name of the current heatmap}

\item{rel_width}{a numeric value decide the relative width of the heatmap
compared to the ggtree plot}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{linejoin}{Line join style (round, mitre, bevel).}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\value{
geom layer
}
\description{
Create heatmaps
}
\author{
Ruizhu Huang
}
