% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_zoomClade.R
\name{geom_th_zoomclade}
\alias{geom_th_zoomclade}
\title{Scale clades}
\usage{
geom_th_zoomclade(node = NULL, scale_x = 1, scale_y = 1)
}
\arguments{
\item{node}{a vector of node}

\item{scale_x}{A numeric vector to scale clades horizontally. It should have
length equal to 1 or the same length as \code{node}.}

\item{scale_y}{A numeric vector to scale clades vertically. It should have
length equal to 1 or the same length as \code{node}.}
}
\value{
a layer
}
\description{
This is to zoom in specific clades.
}
\examples{

library(ggtree)
set.seed(1)
rt <- ape::rtree(200)

ggtree(rt, branch.length = "none") +
    geom_th_zoomclade(node = c(280, 320), scale_y = 8,
    scale_x = c(2, 1.75)) +
    geom_hilight(node = 280, fill = "red", alpha = 0.5) +
    geom_hilight(node = 320, fill = "blue", alpha = 0.5)
}
\author{
Ruizhu Huang
}
