% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_th_axis.R
\name{geom_THaxis}
\alias{geom_THaxis}
\title{ggTHaxis}
\usage{
geom_THaxis(
  mapping = NULL,
  th_data = NULL,
  geom = Geom,
  axis_ticks_length = NULL,
  axis_location = NULL,
  axis_direction = "h",
  draw = NULL,
  ...,
  na.rm = FALSE,
  nudge_x = 0,
  nudge_y = 0,
  name = NULL
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by
\code{\link[ggplot2]{aes}}, which depends on \code{geom}.}

\item{th_data}{a data frame. It should include one column named as
\code{rowLab} to store the row name of the heatmap when \code{side} is
\strong{left} or \strong{right}; otherwise, it should include one column
named as \code{colLab} to store the column name of the heatmap when
\code{side} is \strong{top} or \strong{bottom}.}

\item{geom}{\code{GeomSegment} or \code{GeomText}.}

\item{axis_ticks_length}{a number to decide the lenght of ticks.}

\item{axis_location}{"right" or "left" when \code{axis_direction = "v"};
"top" or "bottom" when \code{axis_direction = "h"}}

\item{axis_direction}{either "h" for horizontal or "v" for vertical}

\item{draw}{"border", "axis" or "axislabel"}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{geom_text}} or
\code{\link[ggplot2]{geom_segment}} for \code{geom = GeomSegment} or
\code{geom = GeomText}.}

\item{na.rm}{If FALSE, the default, missing values are removed with a
warning. If TRUE, missing values are silently removed.}

\item{nudge_x, nudge_y}{Horizontal and vertical adjustment to nudge labels or
lines. Cannot be jointly specified with \code{position}.}

\item{name}{a name to select heatmap}
}
\value{
ggTHaxis (a geom layer)
}
\description{
ggTHaxis
}
\author{
Ruizhu Huang
}
