% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheat.R
\name{ggheat}
\alias{ggheat}
\title{create a heatmap}
\usage{
ggheat(
  data,
  show_rownames = TRUE,
  rownames_position = "left",
  rownames_color = "black",
  rownames_angle = 0,
  rownames_nudge_x = 0,
  rownames_nudge_y = 0,
  rownames_size = 4,
  rownames_hjust = 0.5,
  show_colnames = TRUE,
  colnames_position = "top",
  colnames_color = "black",
  colnames_angle = 0,
  colnames_nudge_x = 0,
  colnames_nudge_y = 0,
  colnames_size = 4,
  colnames_hjust = 0.5
)
}
\arguments{
\item{data}{A matrix or data frame}

\item{show_rownames}{TRUE or FALSE. This specifies whether row names should
be displayed. The default is TRUE.}

\item{rownames_position}{"right" or "left". The default is "left".}

\item{rownames_color}{A color on row names. The default is "black".}

\item{rownames_angle}{A numeric value. The angle of row names. The default is
0.}

\item{rownames_nudge_x}{A numeric value to nudge row names by (on x-axis).
The defaut is 0.}

\item{rownames_nudge_y}{A numeric value to nudge row names by (on y-axis).
The defaut is 0.}

\item{rownames_size}{A numeric value to specify the font size of row names.}

\item{rownames_hjust}{The hjust for row names: 0 (left aligned); 0.5
(centered); 1 (right aligned).}

\item{show_colnames}{TRUE or FALSE. This specifies whether column names should
be displayed. The default is TRUE.}

\item{colnames_position}{"top" or "bottom". The default is "top".}

\item{colnames_color}{A color on column names. The default is "black".}

\item{colnames_angle}{A numeric value. The angle of column names. The default is
0.}

\item{colnames_nudge_x}{A numeric value to nudge column names by (on x-axis).
The defaut is 0.}

\item{colnames_nudge_y}{A numeric value to nudge column names by (on y-axis).
The defaut is 0.}

\item{colnames_size}{A numeric value to specify the font size of column names.}

\item{colnames_hjust}{The hjust for column names: 0 (left aligned); 0.5
(centered); 1 (right aligned).}
}
\value{
heatmap
}
\description{
This creates a heatmap from a matrix
}
\examples{

# matrix
set.seed(2020-05-04)
dd <- matrix(rnorm(20), ncol=5)
rownames(dd) <- paste0('r', 1:4)
colnames(dd) <- paste0('c', 1:5)
ggheat(data = dd)
}
\author{
Ruizhu Huang
}
