% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtree_th.R
\name{ggtree_th}
\alias{ggtree_th}
\title{The tree visualization}
\usage{
ggtree_th(
  tr,
  mapping = NULL,
  layout = "rectangular",
  open.angle = 0,
  mrsd = NULL,
  as.Date = FALSE,
  yscale = "none",
  yscale_mapping = NULL,
  ladderize = TRUE,
  right = FALSE,
  branch.length = "branch.length",
  root.position = 0,
  node = NULL,
  scale_x = 1,
  scale_y = 1,
  ...
)
}
\arguments{
\item{tr}{phylo object}

\item{mapping}{aesthetic mapping}

\item{layout}{one of 'rectangular', 'dendrogram', 'slanted', 'ellipse', 'roundrect',
'fan', 'circular', 'inward_circular', 'radial', 'equal_angle', 'daylight' or 'ape'}

\item{open.angle}{open angle, only for 'fan' layout}

\item{mrsd}{most recent sampling date}

\item{as.Date}{logical whether using Date class in time tree}

\item{yscale}{y scale}

\item{yscale_mapping}{yscale mapping for category variable}

\item{ladderize}{logical (default \code{TRUE}). Should the tree be re-organized to have a 'ladder'
aspect?}

\item{right}{logical. If \code{ladderize = TRUE}, should the ladder have the smallest clade on the
right-hand side? See \code{\link[ape:ladderize]{ape::ladderize()}} for more information.}

\item{branch.length}{variable for scaling branch, if 'none' draw cladogram}

\item{root.position}{position of the root node (default = 0)}

\item{node}{Nodes to specify clades for zoom-in}

\item{scale_x}{A numeric vector to scale clades horizontally. It should have
length equal to 1 or the same length as \code{node}.}

\item{scale_y}{A numeric vector to scale clades vertically. It should have
length equal to 1 or the same length as \code{node}.}

\item{...}{additional parameter

some dot arguments:
\itemize{
\item \code{nsplit} integer, the number of branch blocks divided when 'continuous' is not "none", default is 200.
}}
}
\value{
ggtree
}
\description{
This
}
\examples{
library(ggtree)
set.seed(1)
rt <- ape::rtree(200)

#ggtree_th(rt, branch.length = "none", node = c(280, 320),
#                  scale_y = 8, scale_x = c(2, 1.75)) +
#    geom_hilight(node = 280, fill = "red", alpha = 0.5) +
#    geom_hilight(node = 320, fill = "blue", alpha = 0.5)
}
\author{
Ruizhu Huang
}
