% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatdf.R
\name{heatdf}
\alias{heatdf}
\title{convert data}
\usage{
heatdf(data, fig_rowtree = NULL, fig_coltree = NULL, w = 1, h = 1)
}
\arguments{
\item{data}{a matrix}

\item{fig_rowtree}{a gg object created by \code{\link[ggtree]{ggtree}} on row tree}

\item{fig_coltree}{a gg object created by \code{\link[ggtree]{ggtree}} on column tree}

\item{w}{a numeric value to specify column width.}

\item{h}{a numeric value to specify row height.}
}
\value{
a data frame
}
\description{
convert data
}
