% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Oral}
\alias{Oral}
\title{16S rRNA data (Oral site)}
\format{
A list of datasets:
\describe{
  \item{count_table}{A matrix storing counts of 80 OTUs in 124 samples}
  \item{taxo_table}{The taxonomic table of the 80 OTUs in count_table}
  \item{meta_sample}{The metadata of the 124 samples in count_table}
  \item{phylo_tree}{A phylogenetic tree of the 80 OTUs}
  \item{taxo_tree}{A taxonomic tree of the 80 OTUs}
}
}
\source{
The original data is from \code{\link[HMP16SData]{V35}}. The code
  used to generate the 'Oral' data is available as the script file
  \code{Oral.R} in the folder \code{data-raw} of this package.
}
\usage{
Oral
}
\description{
A list of datasets containing HMP 16S rRNA data collected from the oral site.
It includes 80 OTUs in 124 samples. The OTUs are the top 80 highly variable
OTUs in samples. Note: Counts of OTUs are in cpm rather than the raw count.
}
\keyword{datasets}
