% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_row.R
\name{rowAnnotate}
\alias{rowAnnotate}
\title{add annotation on rows of heatmap}
\usage{
rowAnnotate(plot, anno_plot, side = "left", width = 0.1, name = NULL)
}
\arguments{
\item{plot}{the main heatmap}

\item{anno_plot}{A annotation plot}

\item{side}{"left" or "right"}

\item{width}{0.1}

\item{name}{NULL}
}
\description{
\code{anno_row} annotates rows of heatmap, either on the left or right side.
}
