% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_th_text.R
\docType{data}
\name{StatHeatText}
\alias{StatHeatText}
\title{StatHeatText}
\format{
An object of class \code{StatHeatText} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
StatHeatText
}
\description{
StatHeatText
}
\keyword{datasets}
