% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatTH.R
\docType{data}
\name{StatTH}
\alias{StatTH}
\title{StatTH}
\format{
An object of class \code{StatTH} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
StatTH
}
\description{
StatTH
}
\keyword{datasets}
