% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stool}
\alias{stool}
\title{The stool data}
\format{
A list of datasets:
\describe{
  \item{norm_count}{A matrix with 464 rows and 30 columns. Each row
  represents a leaf of \code{tree}, and each column represent a sample.
  Samples are collected at different time points (0, 4, or 13 months) on
  babies that were born differently (c-section or vaginal). For example,
  \code{0 M:c_section_1} means that the sample was collected on 0 month
  (\code{0 M}) from baby (\code{_1}) in the c-section (\code{c_section}) group.}
  \item{tree}{A tree with 464 leaves and 463 internal nodes.}
  \item{df_branch}{A data frame has 464 rows and 3 columns. node: the node
  numeric id; Detected: Yes or No; pick: TRUE or FALSE.}
}
}
\source{
The data is generated
randomly with the R script \code{stool.R} in the \code{data-raw} folder of
this package.
}
\usage{
stool
}
\description{
It contains a normalized count table \code{norm_count}, and a tree, and a
data frame \code{df_branch}. Note: this is a processed data only for
visualization here, and we don't suggest to use it for data analysis.
}
\keyword{datasets}
