% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_th_axis.R
\name{geom_THaxis}
\alias{geom_THaxis}
\alias{geom_th_axis}
\title{add the axis and the border
add the axis and the border of an annotation plot}
\usage{
geom_THaxis(
  name = NULL,
  mapping = NULL,
  th_data = NULL,
  axis_ticks_length = NULL,
  axis_location = NULL,
  draw = NULL,
  axis_direction = "h",
  n = 5,
  geom = Geom,
  position = "identity",
  stat = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  nudge_x = 0,
  nudge_y = 0
)

geom_th_axis(
  th_data = NULL,
  n = 5,
  axis_ticks_length = NULL,
  axis_color = "black",
  axis_label_color = "black",
  axis_label_size = 3,
  axis_label_nudge_x = 0,
  axis_label_nudge_y = 0,
  axis_direction = "h",
  axis_location = NULL,
  border_color = "grey",
  border_size = 0.5
)
}
\arguments{
\item{th_data}{a data frame. To draw a vertical axis, it requires columns
\code{axis_minX}, \code{axis_maxX} and \code{axis_minY}, where \code{axis_minX}
and \code{axis_maxX} decide the border limit on the x-dim. The axis is
either on \code{axis_minX} or \code{axis_maxX} with ticks decided by the
\code{axis_minY}. Note, we use \link[base]{pretty} to decide breaks on as
\code{axis_minY} ticks. Similarly, to draw a horizontal axis, it requires
columns \code{axis_minX}, \code{axis_minY}, \code{axis_maxY}.}

\item{axis_ticks_length}{a number to decide the lenght of ticks.}

\item{axis_direction}{either "h" for horizontal or "v" for vertical}

\item{n}{a number to decide the number of ticks. See \link[base]{pretty}.}

\item{axis_color}{the color of the axis}

\item{axis_label_color}{the color of axis labels}

\item{axis_label_size}{the size of axis labels}

\item{axis_label_nudge_x}{adjust the location of labels along x-axis}

\item{axis_label_nudge_y}{adjust the location of labels along y-axis}

\item{border_color}{the color of the rectangular border}

\item{border_size}{the size of the border line}
}
\value{
ggTHaxis (a geom layer)
}
\description{
add the axis and the border
add the axis and the border of an annotation plot
}
\examples{
library(ggplot2)
df_v <- data.frame(axis_minX = rep(1, 10),
                   axis_maxX = rep(10, 10),
                   axis_minY = 1:10)
ggplot() +
    geom_th_axis(th_data = df_v, axis_direction = "v", n = 5)


df_h <- data.frame(axis_minX = 1:10,
                   axis_minY = rep(1, 10),
                   axis_maxY = rep(10, 10))


ggplot() +
    geom_th_axis(th_data = df_h, axis_direction = "h", n = 5)
}
\author{
Ruizhu Huang

Ruizhu Huang
}
