% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_th_bar.R
\name{geom_th_bar}
\alias{geom_th_bar}
\title{add barplots}
\usage{
geom_th_bar(
  name = NULL,
  th_data = NULL,
  mapping = NULL,
  side = "right",
  gap = 1,
  rel_width = 1,
  stat = "identity",
  ...
)
}
\arguments{
\item{name}{a name to select heatmap}

\item{th_data}{a data frame. It should include one column named as
\code{rowLab} to store the row name of the heatmap when \code{side} is
\strong{left} or \strong{right}; otherwise, it should include one column
named as \code{colLab} to store the column name of the heatmap when
\code{side} is \strong{top} or \strong{bottom}.}

\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{side}{a character to specify where to put the barplot. It is selected
from \strong{left}, \strong{right}, \strong{top} or \strong{bottom}.}

\item{gap}{a numeric value to specify the gap between the selected heatmap
and the barplot}

\item{rel_width}{a numeric value decide the relative width of the barplot
compared to the selected heatmap}

\item{stat}{Please see \code{stat} of \code{\link[ggplot2]{geom_bar}}.}

\item{...}{More arguements accepted by \code{\link[ggplot2]{geom_bar}}.}
}
\value{
geom layer
}
\description{
add barplots
}
\author{
Ruizhu Huang
}
