% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_th_border.R
\name{geom_th_border}
\alias{geom_th_border}
\title{add the border line}
\usage{
geom_th_border(
  name,
  side,
  gap = 0.5,
  extend_y = c(0, 0),
  extend_x = c(0, 0),
  label_color = "black",
  label_size = 4,
  label_offset_x = 0,
  label_offset_y = 0,
  label = "",
  ...
)
}
\arguments{
\item{name}{the name of the heatmap to add row or column labels.}

\item{side}{a character value selected from \strong{left}, \strong{right},
\strong{top} or \strong{bottom}. \strong{left} and \strong{right} to add
the vertical line; \strong{top} or \strong{bottom} to add the horizontal
line.}

\item{gap}{the gap between the border line and the heatmap}

\item{extend_y}{a numeric vector with two elements. For example, \code{c(1,
2)} would extend the bottom end of the line by 1 and the top end by 2.}

\item{extend_x}{a numeric vector with two elements. For example, \code{c(1,
2)} would extend the left end of the line by 1 and the right end by 2.}

\item{label_color}{the color of the label.}

\item{label_size}{the font size of the label.}

\item{label_offset_x}{a value to shift the label horizontally.}

\item{label_offset_y}{a value to shift the label vertically.}

\item{label}{a character to label the border line}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\value{
geom layer
}
\description{
add a line to separate heatmaps
}
\seealso{
\code{\link{geom_th_borderline}} \code{\link{geom_th_bordertext}}
}
\author{
Ruizhu Huang
}
