% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_th_coltree.R
\name{geom_th_coltree}
\alias{geom_th_coltree}
\title{add the column tree}
\usage{
geom_th_coltree(name = NULL, gap = 0.5, rel_height = 0.1, ...)
}
\arguments{
\item{name}{the name of the heatmap}

\item{gap}{a numeric value to specify the gap between the column tree and the
heatmap}

\item{rel_height}{a numeric value decide the relative height of the column
tree compared to the height of the heatmap}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\value{
geom layer
}
\description{
add the column tree
}
\author{
Ruizhu Huang
}
