% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_th_axis.R
\name{geom_THaxis}
\alias{geom_THaxis}
\title{ggTHaxis}
\usage{
geom_THaxis(
  mapping = NULL,
  th_data = NULL,
  axis_ticks_length = NULL,
  axis_location = NULL,
  axis_direction = "h",
  draw = NULL,
  geom = Geom,
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE,
  nudge_x = 0,
  nudge_y = 0,
  name = NULL
)
}
\arguments{
\item{th_data}{a data frame. It should include one column named as
\code{rowLab} to store the row name of the heatmap when \code{side} is
\strong{left} or \strong{right}; otherwise, it should include one column
named as \code{colLab} to store the column name of the heatmap when
\code{side} is \strong{top} or \strong{bottom}.}

\item{axis_ticks_length}{a number to decide the lenght of ticks.}

\item{axis_location}{"right" or "left" when \code{axis_direction = "v"};
"top" or "bottom" when \code{axis_direction = "h"}}

\item{axis_direction}{either "h" for horizontal or "v" for vertical}

\item{draw}{"border", "axis" or "axislabel"}

\item{name}{a name to select heatmap}

\item{border_color}{the color of the rectangular border}

\item{border_size}{the size of the border line}
}
\value{
ggTHaxis (a geom layer)
}
\description{
ggTHaxis
}
\author{
Ruizhu Huang
}
