% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{The toy data}
\format{
A list of datasets:
\describe{
  \item{toytree}{A tree with 10 leaves and 9 internal nodes.}
  \item{countA}{A matrix has 10 rows and 8 columns}
  \item{countB}{A matrix has 7 rows and 8 columns}
}
}
\source{
The data is generated
randomly with the R script \code{toydata.R} in the \code{data-raw} folder of
this package.
}
\usage{
toydata
}
\description{
It contains a toy tree, and two count matrices. The data is generated
randomly with the R script \code{toydata.R} in the \code{data-raw} folder of
this package. The toy tree has 10 leaves. Each count matrix has 8 columns
representing 8 samples. Each row of the count matrix corresponds to a node of
the tree.
}
\keyword{datasets}
